fix gcc warning
git-svn-id: file:///usr/local/Singular/svn/trunk@12975 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ kernel/int64vec.h
@@ -40,7 +40,7 @@ public:
 #endif
       return v[i];
     }
-  inline const int& operator[](int i) const
+  inline const int64& operator[](int i) const
     {
 #ifndef NDEBUG
       if((i<0)||(i>=row*col))
